+++ src/main/java/com/braintreegateway/TransactionGateway.java
@@ -41,6 +41,8 @@ public class TransactionGateway {
 
     /**
      * Please use gateway.transparentRedirect().confirmTransaction() instead
+     * @param queryString the query string
+     * @return the transaction
      */
     @Deprecated
     public Result<Transaction> confirmTransparentRedirect(String queryString) {
@@ -124,6 +126,7 @@ public class TransactionGateway {
     /**
      * Finds all Transactions that match the query and returns a {@link ResourceCollection}.
      * See: <a href="https://developers.braintreepayments.com/reference/request/transaction/search/java" target="_blank">https://developers.braintreepayments.com/reference/request/transaction/search/java</a>
+     * @param query the search query
      * @return a {@link ResourceCollection} or raises a {@link DownForMaintenanceException}.
      */
     public ResourceCollection<Transaction> search(TransactionSearchRequest query) {
@@ -229,6 +232,7 @@ public class TransactionGateway {
 
     /**
      * Please use gateway.transparentRedirect().url() instead
+     * @return redirect URL
      */
     @Deprecated
     public String transparentRedirectURLForCreate() {