+++ src/test/java/com/braintreegateway/unittest/OAuthAccessRevocationTest.java
@@ -12,10 +12,12 @@ public class OAuthAccessRevocationTest {
     public void assignsMerchantId() {
         String xmlPayload = "<oauth-application-revocation>" +
             "<merchant-id>abc123def</merchant-id>" +
+            "<oauth-application-client-id>uvw789xyz</oauth-application-client-id>" +
             "</oauth-application-revocation>";
         NodeWrapper node = NodeWrapperFactory.instance.create(xmlPayload);
         OAuthAccessRevocation revocation = new OAuthAccessRevocation(node);
 
         assertEquals(revocation.getMerchantId(), "abc123def");
+        assertEquals(revocation.getOauthApplicationClientId(), "uvw789xyz");
     }
 }