+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.81.0
+* Allow payee ID to be passed in options params for transaction create
+* Add `getMerchantId` to `ConnectedMerchantStatusTransitioned` and `ConnectedMerchantPayPalStatusChanged`, and `getOauthApplicationClientId` to OAuthAccessRevocation webhooks
+
 ## 2.80.0
 * Add support for US Bank Account verifications API